backend/pyproject.toml
[project]
name = "aura-backend"
version = "poetry shel"
description = ""
authors = [
    {name = "cTxdeveloper",email = "ttok17723@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "beanie (>=1.30.0,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
